preload task status
